requirements.txt
praw
python-dotenv 
openai